requirements.txt
torch
torchvision
tqdm
torchmetrics
torchinfo
Pillow
flask